pyproject.toml
[project]
name = "lumos"
version = "0.1.0"
description = "Agent identity and capability authorization for MCP"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.30.0",
  "cryptography>=42.0.0",
  "fastapi>=0.115.0",
  "PyJWT>=2.8.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
